requirements.txt
streamlit==1.31.0
supabase==2.3.4
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
numpy==1.26.3
sqlalchemy
psycopg2-binary
